+++ apps/backoffice-v2/src/pages/Entity/hooks/useEntity/utils.ts
@@ -1,6 +1,7 @@
 import { TDropdownOption } from '../../components/EditableDetails/types';
 import { AnyArray } from '../../../../common/types';
 import { TDocument } from '@ballerine/common';
+import { TWorkflowById } from '../../../../domains/workflows/fetchers';
 
 export const convertSnakeCaseToTitleCase = (input: string): string =>
   input
@@ -43,7 +44,7 @@ export const composePickableCategoryType = (
     const category = document.category;
     if (category) {
       documentCategoryDropdownOptions.push({
-        value: category as string,
+        value: category,
         label: convertSnakeCaseToTitleCase(category),
       });
     }
@@ -51,8 +52,8 @@ export const composePickableCategoryType = (
     if (type) {
       documentTypesDropdownOptions.push({
         dependantOn: 'category',
-        dependantValue: category as string,
-        value: type as string,
+        dependantValue: category,
+        value: type,
         label: convertSnakeCaseToTitleCase(type),
       });
     }
@@ -66,11 +67,11 @@ export const composePickableCategoryType = (
     ...composeDataFormCell('type', typeDropdownOptions, typeValue),
   };
 };
-export const isExistingSchemaForDocument = documentsSchemas => {
-  return documentsSchemas && documentsSchemas.length > 0;
+export const isExistingSchemaForDocument = (documentsSchemas: Array<TDocument>) => {
+  return documentsSchemas?.length > 0;
 };
 
-export const extractCountryCodeFromWorkflow = workflow => {
+export const extractCountryCodeFromWorkflow = (workflow: TWorkflowById) => {
   return workflow?.context?.documents?.find(document => {
     return !!document?.issuer?.country;
   })?.issuer?.country;