+++ examples/kyb-app/src/common/providers/ViewStateProvider/hooks/useViewsDataRepository/types.ts
@@ -1,0 +1,8 @@
+import { AnyObject } from '@ballerine/ui';
+
+export interface ViewsData {
+  currentView: string;
+  shared: AnyObject;
+  flowData: AnyObject;
+  completionMap: Record<string, boolean>;
+}