+++ examples/kyb-app/src/common/providers/ViewStateProvider/types.ts
@@ -1,22 +1,39 @@
+import { InputsWarnings } from '@app/common/components/organisms/DynamicForm';
+import { IStep, StepMetadata } from '@app/common/hooks/useStepper';
 import { AnyObject } from '@ballerine/ui';
-import { MachineConfig } from 'xstate';
 
 export type SchemaBase = { states: AnyObject };
 
 export type SchemaStates<TSchema extends SchemaBase> = keyof TSchema['states'];
 
-export interface ViewStateContext<
-  TGlobalContext = AnyObject,
-  TViewContext = AnyObject,
-  TStates = unknown,
-> {
-  next: () => void;
-  prev: () => void;
-  saveAndPerformTransition: (data: object, shared?: object) => void;
-  send: (type: TStates, formData?: object) => void;
-  state: string;
-  stateData: TViewContext;
-  context: TGlobalContext;
+export interface InitialContext<TSchema extends SchemaBase> {
+  state: SchemaStates<TSchema>;
 }
 
-export type ViewStateSchema = MachineConfig<any, any, any>;
+export interface StepperParams {
+  currentStep: number;
+  totalSteps: number;
+}
+
+export interface ViewStateContext<TGlobalContext = AnyObject> {
+  next: () => void;
+  prev: () => void;
+  update: (data: object, shared?: object) => Promise<object>;
+  save: <T>(data: T, shared?: object) => Promise<TGlobalContext>;
+  saveAndPerformTransition: <T>(data: T, shared?: object) => Promise<TGlobalContext>;
+  finish: (context: TGlobalContext) => void;
+  context: TGlobalContext;
+  state: string | number;
+  steps: IStep[];
+  stepper: StepperParams;
+  warnings?: InputsWarnings;
+  isFinished: boolean;
+}
+
+export interface View {
+  label: string;
+  key: string;
+  active?: boolean;
+  Component: React.ComponentType;
+  meta?: StepMetadata;
+}