+++ examples/kyb-app/src/components/atoms/Stepper/components/atoms/indicators/Idle.tsx
@@ -1,0 +1,12 @@
+import {
+  BaseIndicator,
+  BaseIndicatorProps,
+} from '@app/components/atoms/Stepper/components/atoms/BaseIndicator';
+
+interface Props {
+  indicatorProps?: BaseIndicatorProps;
+}
+
+export const Idle = ({ indicatorProps }: Props) => {
+  return <BaseIndicator {...indicatorProps} />;
+};