+++ examples/kyb-app/src/components/atoms/Stepper/hooks/useStepperContext/useStepperContext.ts
@@ -1,0 +1,4 @@
+import { stepperContext } from '@app/components/atoms/Stepper/stepper.context';
+import { useContext } from 'react';
+
+export const useStepperContext = () => useContext(stepperContext);