+++ examples/kyb-app/src/components/organisms/KYBView/views/PersonalInformationView/PersonalInformationView.tsx
@@ -1,35 +1,49 @@
 import { DynamicForm } from '@app/common/components/organisms/DynamicForm';
 import { useViewState } from '@app/common/providers/ViewStateProvider';
 import { AppShell } from '@app/components/layouts/AppShell';
+import { ViewHeader } from '@app/components/organisms/KYBView/components/ViewHeader';
 import { useCreateEndUserMutation } from '@app/components/organisms/KYBView/hooks/useCreateEndUserMutation';
-import { PersonalInformationContext } from '@app/components/organisms/KYBView/types';
 import { formSchema } from '@app/components/organisms/KYBView/views/PersonalInformationView/form.schema';
+import {
+  PersonalInformationContext,
+  WorkflowFlowData,
+} from '@app/domains/workflows/flow-data.type';
 import { useCallback } from 'react';
 
 export const PersonalInformationView = () => {
-  const { saveAndPerformTransition } = useViewState();
+  const { context, state, saveAndPerformTransition } = useViewState<WorkflowFlowData>();
   const { createUserAsync } = useCreateEndUserMutation();
 
   const handleSubmit = useCallback(
     (values: PersonalInformationContext) => {
+      if (context.shared.endUserId && context.shared.businessId) {
+        void saveAndPerformTransition(values, {
+          endUserId: context.shared.endUserId,
+          businessId: context.shared.businessId,
+        });
+
+        return;
+      }
+
       createUserAsync(values)
         .then(result => {
-          saveAndPerformTransition(values, {
+          void saveAndPerformTransition(values, {
             endUserId: result.endUserId,
             businessId: result.businessId,
           });
         })
         .catch(e => {
-          console.log('failed to create user', e);
+          console.log('Failed to create user', e);
         });
     },
-    [saveAndPerformTransition, createUserAsync],
+    [context, saveAndPerformTransition, createUserAsync],
   );
 
   return (
-    <AppShell.FormContainer>
-      <DynamicForm
+    <AppShell.FormContainer header={<ViewHeader progressBar={false} />}>
+      <DynamicForm<PersonalInformationContext>
         className="max-w-[384px]"
+        formData={context.flowData[state] as PersonalInformationContext}
         uiSchema={{
           'ui:options': {
             submitButtonOptions: {