+++ examples/kyb-app/src/domains/workflows/serialize/run-and-start-workflow.serialize.ts
@@ -1,6 +1,6 @@
-import { RunWorkflowDto, TRunWorkflowDto } from '@app/domains/workflows/types';
+import { TRunWorkflowDto, WorkflowUpdatePayload } from '@app/domains/workflows/types';
 
-export const runAndStartWorkflowSerialize = (data: RunWorkflowDto): TRunWorkflowDto => {
+export const serializeWorkflowUpdatePayload = (data: WorkflowUpdatePayload): TRunWorkflowDto => {
   const payload: TRunWorkflowDto = {
     workflowId: data.workflowId,
     context: {
@@ -12,13 +12,21 @@ export const runAndStartWorkflowSerialize = (data: RunWorkflowDto): TRunWorkflow
           website: data.entity.website,
           registrationNumber: data.entity.registrationNumber,
           companyName: data.entity.companyName,
+          countryOfIncorporation: data.entity.country, // TODO: this needs to be extracted from address
           address: {
             text: data.entity.address,
           },
           additionalInfo: {
-            customerCompany: 'Ballerine',
-            companyName: data.entity.companyName,
-            ubos: data.entity.ubos,
+            mainRepresentative: data.entity.mainRepresentative,
+            ubos: data.entity.ubos.map(ubo => {
+              ubo.entity.data.additionalInfo = {
+                ...(ubo.entity.data.additionalInfo || {}),
+                companyName: data.entity.companyName,
+                customerCompany: 'PayLink',
+              };
+
+              return ubo;
+            }),
           },
         },
       },
@@ -28,6 +36,7 @@ export const runAndStartWorkflowSerialize = (data: RunWorkflowDto): TRunWorkflow
         issuer: {
           country,
         },
+        decision: { status: '', revisionReason: '', rejectionReason: '' },
         pages: pages.map(({ fileId }) => ({ ballerineFileId: fileId })),
         properies: {},
         version: '1',