+++ packages/ui/src/common/types.ts
@@ -1,4 +1,5 @@
 import { ReactNode } from 'react';
 
-export type AnyChildren = ReactNode | ReactNode[];
+export type ChildrenList = ReactNode[];
+export type AnyChildren = ReactNode | ChildrenList;
 export type AnyObject = Record<PropertyKey, any>;