+++ packages/workflow-core/src/lib/utils/context-transformers/helpers-transformer.ts
@@ -31,20 +31,22 @@ export class HelpersTransformer extends BaseContextTransformer {
     return context;
   }
 
-  async regex(attribute: string, value: string) {
+  regex(attribute: string, value: string) {
     const result = attribute.match(new RegExp(value));
 
-    return Promise.resolve(result ? result[0] : null);
+    return result ? result[0] : null;
   }
 
   async imageUrlToBase64(url: string, _value: string) {
     const response = await fetch(url);
+    const imageType = response?.headers?.get('Content-Type')?.split(';')?.[0] || 'image/png';
     const arrayBuffer = await response.arrayBuffer();
     const buffer = Buffer.from(arrayBuffer);
-    const Base64Image = buffer.toString('base64');
-    const base64Prefix = 'data:image/png;base64,';
 
-    return base64Prefix + Base64Image;
+    const base64Image = buffer.toString('base64');
+    const base64Prefix = `data:${imageType};base64,`;
+
+    return base64Prefix + base64Image;
   }
 
   getNestedProperty(record: Record<string, any>, path: Array<string>) {