+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -154,7 +154,6 @@ export class WorkflowService {
   }
 
   async getWorkflowRuntimeWithChildrenDataById(
-    // daniel
     id: string,
     args?: Parameters<WorkflowRuntimeDataRepository['findById']>[1],
   ) {
@@ -164,13 +163,20 @@ export class WorkflowService {
   }
 
   async getWorkflowByIdWithRelations(
-    //
     id: string,
     args?: Parameters<WorkflowRuntimeDataRepository['findById']>[1],
   ) {
+    const allEntities = { endUser: true, business: true };
+    const childWorkflowSelectArgs = {
+      select: { ...args?.select, ...allEntities },
+      include: args?.include,
+    };
     const workflow = (await this.workflowRuntimeDataRepository.findById(id, {
       ...args,
-      select: { ...(args?.select || {}), childWorkflowsRuntimeData: { ...args } },
+      select: {
+        ...(args?.select || {}),
+        childWorkflowsRuntimeData: { ...childWorkflowSelectArgs },
+      },
     })) as TWorkflowWithRelations;
 
     return this.formatWorkflow(workflow);
@@ -190,7 +196,7 @@ export class WorkflowService {
         definitionType: workflow.workflowDefinition.definitionType,
         workflowContext: {
           machineContext: workflow.context,
-          state: workflow.state ?? workflow.workflowDefinition.definition.initial,
+          state: workflow.state ?? workflow.workflowDefinition.definition?.initial,
         },
       });
 
@@ -244,12 +250,24 @@ export class WorkflowService {
       })
     )[0];
 
+    const parentWorkflowRuntime = await this.getWorkflowRuntimeDataById(
+      childPluginConfig.parentWorkflowRuntimeId,
+    );
+
     if (childWorkflow) {
-      await this.persistChildWorkflowToParent(
-        childWorkflow.workflowRuntimeData,
-        childWorkflow.workflowDefinition,
-        false,
+      const contextToPersist = {
+        [childWorkflow.workflowRuntimeData.id]: {
+          entityId: childWorkflow.workflowRuntimeData.context.entity.id,
+          status: childWorkflow.workflowRuntimeData.status || 'active',
+        },
+      };
+      const parentContext = this.composeContextWithChildResponse(
+        parentWorkflowRuntime.context,
+        childWorkflow.workflowDefinition.id,
+        contextToPersist,
       );
+
+      await this.updateWorkflowRuntimeData(parentWorkflowRuntime.id, { context: parentContext });
     }
 
     return childWorkflow;
@@ -490,6 +508,72 @@ export class WorkflowService {
     return await this.workflowDefinitionRepository.findMany({ ...args, select });
   }
 
+  async updateDecisionAndSendEvent({
+    id,
+    name,
+    reason,
+  }: {
+    id: string;
+    name: string;
+    reason?: string;
+  }) {
+    const runtimeData = await this.workflowRuntimeDataRepository.findById(id);
+    // `name` is always `approve` and not `approved` etc.
+    const Status = {
+      approve: 'approved',
+      reject: 'rejected',
+      revision: 'revision',
+    } as const;
+    const status = Status[name as keyof typeof Status];
+    const decision = (() => {
+      if (status === 'approved') {
+        return {
+          revisionReason: null,
+          rejectionReason: null,
+        };
+      }
+
+      if (status === 'rejected') {
+        return {
+          revisionReason: null,
+          rejectionReason: reason,
+        };
+      }
+
+      if (status === 'revision') {
+        return {
+          revisionReason: reason,
+          rejectionReason: null,
+        };
+      }
+
+      throw new BadRequestException(`Invalid decision status: ${status}`);
+    })();
+    const documentsWithDecision = runtimeData?.context?.documents?.map(
+      (document: DefaultContextSchema['documents'][number]) => ({
+        ...document,
+        decision: {
+          ...document?.decision,
+          ...decision,
+          status,
+        },
+      }),
+    );
+    const updatedWorkflow = await this.updateWorkflowRuntimeData(id, {
+      context: {
+        ...runtimeData?.context,
+        documents: documentsWithDecision,
+      },
+    });
+
+    await this.event({
+      id,
+      name,
+    });
+
+    return updatedWorkflow;
+  }
+
   async updateWorkflowRuntimeData(workflowRuntimeId: string, data: WorkflowDefinitionUpdateInput) {
     const runtimeData = await this.workflowRuntimeDataRepository.findById(workflowRuntimeId);
     const workflowDef = await this.workflowDefinitionRepository.findById(
@@ -568,6 +652,13 @@ export class WorkflowService {
 
     if (isResolved) {
       this.logger.log('Workflow resolved', { id: workflowRuntimeId });
+
+      this.workflowEventEmitter.emit('workflow.completed', {
+        runtimeData,
+        state: currentState ?? runtimeData.state,
+        entityId: runtimeData.businessId || runtimeData.endUserId,
+        correlationId,
+      });
     }
 
     const documentToRevise = data.context?.documents?.find(
@@ -1094,7 +1185,9 @@ export class WorkflowService {
       extensions: workflowDefinition.extensions,
       invokeChildWorkflowAction: async (childPluginConfiguration: ChildPluginCallbackOutput) => {
         const runnableChildWorkflow = await this.persistChildEvent(childPluginConfiguration);
+
         if (runnableChildWorkflow && childPluginConfiguration.initOptions.event) {
+          // TODO: Review the issue if return child workflow id for parent and not "send event"
           await this.event({
             id: runnableChildWorkflow.workflowRuntimeData.id,
             name: childPluginConfiguration.initOptions.event,
@@ -1126,8 +1219,13 @@ export class WorkflowService {
       status: isFinal ? 'completed' : workflowRuntimeData.status,
     });
 
-    if (isFinal && workflowRuntimeData.parentRuntimeDataId) {
-      await this.persistChildWorkflowToParent(workflowRuntimeData, workflowDefinition, isFinal);
+    if (workflowRuntimeData.parentRuntimeDataId) {
+      await this.persistChildWorkflowToParent(
+        workflowRuntimeData,
+        workflowDefinition,
+        isFinal,
+        currentState,
+      );
     }
 
     this.workflowEventEmitter.emit('workflow.state.changed', {
@@ -1172,6 +1270,7 @@ export class WorkflowService {
     workflowRuntimeData: WorkflowRuntimeData,
     workflowDefinition: WorkflowDefinition,
     isFinal: boolean,
+    childRuntimeState?: string,
   ) {
     const parentWorkflowRuntime = await this.getWorkflowRuntimeWithChildrenDataById(
       workflowRuntimeData.parentRuntimeDataId,
@@ -1195,12 +1294,20 @@ export class WorkflowService {
       childWorkflow =>
         childWorkflow.workflowDefinitionId === workflowRuntimeData.workflowDefinitionId,
     );
+    const isPersistableState =
+      !!(
+        childRuntimeState &&
+        childWorkflowCallback.persistenceStates &&
+        childWorkflowCallback.persistenceStates.includes(childRuntimeState)
+      ) || isFinal;
+    if (!isPersistableState) return;
+
     const parentContext = await this.generateParentContextWithInjectedChildContext(
       childrenOfSameDefinition,
       childWorkflowCallback.transformers,
       parentWorkflowRuntime,
       workflowDefinition,
-      isFinal,
+      isPersistableState,
     );
 
     await this.updateWorkflowRuntimeData(parentWorkflowRuntime.id, { context: parentContext });
@@ -1208,7 +1315,7 @@ export class WorkflowService {
     if (
       childWorkflowCallback.deliverEvent &&
       parentWorkflowRuntime.status !== 'completed' &&
-      isFinal
+      isPersistableState
     ) {
       await this.event({
         id: parentWorkflowRuntime.id,
@@ -1222,9 +1329,9 @@ export class WorkflowService {
     transformers: ChildWorkflowCallback['transformers'],
     parentWorkflowRuntime: WorkflowRuntimeData,
     workflowDefinition: WorkflowDefinition,
-    isFinal: boolean,
+    isPersistableToParent: boolean,
   ) {
-    if (!isFinal)
+    if (!isPersistableToParent)
       return this.composeContextWithChildResponse(
         parentWorkflowRuntime.context,
         workflowDefinition.id,