+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.4.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.14
+  - @ballerine/workflow-core@0.4.24
+
+## 0.4.16
+
+### Patch Changes
+
+- Upgrade packages
+- Updated dependencies
+  - @ballerine/common@0.5.13
+  - @ballerine/workflow-core@0.4.23
+
+## 0.4.15
+
+### Patch Changes
+
+- Update versions
+- Updated dependencies
+  - @ballerine/workflow-core@0.4.22
+  - @ballerine/common@0.5.12
+
+## 0.4.14
+
+### Patch Changes
+
+- Packages update
+- Updated dependencies
+  - @ballerine/workflow-core@0.4.21
+  - @ballerine/common@0.5.11
+
+## 0.4.13
+
+### Patch Changes
+
+- Update packages
+- Updated dependencies
+  - @ballerine/workflow-core@0.4.20
+  - @ballerine/common@0.5.10
+
+## 0.4.12
+
+### Patch Changes
+
+- New UI package
+- Updated dependencies
+  - @ballerine/workflow-core@0.4.19
+  - @ballerine/common@0.5.9
+
 ## 0.4.11
 
 ### Patch Changes